lxcinstaller: add installer for LXC containers
This adds a LXCInstaller class that will bootstrap the operating system
trees in the LXC containers.

It has an installer phase that runs within the host network and
filesystem tree, but with the uid/gid mapping already set up within the user
namespace (in case it's configured).

This allows fetching the packages from the internet and unpacking the
files with the ownership that makes sense in the installed container.

The actual distribution support will be added in follow-up commits.